Systeminfo plugin: add systeminfo model

+++ plugins/systeminfo/systeminfothread.cpp
@@ -1,8 +1,10 @@
 #include "systeminfothread.h"
+#include "networkdevicemodel.h"
 #include <QFile>
 #include <QNetworkInterface>
 
-SysteminfoThread::SysteminfoThread(QObject *parent) : QThread(parent)
+SysteminfoThread::SysteminfoThread(SystemInfoModel *model, QObject *parent) :
+    m_model(model)
 {
     m_rx = nullptr;
     m_tx = nullptr;
@@ -14,25 +16,44 @@ void SysteminfoThread::run()
     quint64 old_rx;
 
     for (;;) {
+        QString device;
         foreach (const QNetworkInterface &interface, QNetworkInterface::allInterfaces()) {
-            if (interface.flags().testFlag(QNetworkInterface::IsUp) && !interface.flags().testFlag(QNetworkInterface::IsLoopBack)) {
-                //                foreach (QNetworkAddressEntry entry, interface.addressEntries()) {
-                //                    if ( interface.hardwareAddress() != "00:00:00:00:00:00" && entry.ip().toString().contains(".")) {
-                //                        // qDebug() << interface.name() + " "+ entry.ip().toString() +" " + interface.hardwareAddress();
-                //                        m_deviceName = interface.name();
-                //                    }
-                //                }
-                if (m_rx) {
-                    m_rx->deleteLater();
-                    m_rx = nullptr;
-                }
-                if (m_tx) {
-                    m_tx->deleteLater();
-                    m_tx = nullptr;
-                }
-                m_rx = new QFile("/sys/class/net/" + interface.name() + "/statistics/rx_bytes");
-                m_tx = new QFile("/sys/class/net/" + interface.name() + "/statistics/tx_bytes");
+            if (!device.isEmpty())
                 break;
+
+            if (interface.flags().testFlag(QNetworkInterface::IsUp) && !interface.flags().testFlag(QNetworkInterface::IsLoopBack)) {
+                foreach (QNetworkAddressEntry entry, interface.addressEntries()) {
+                    if ( interface.hardwareAddress() != "00:00:00:00:00:00" && entry.ip().toString().contains(".")) {
+//                         qDebug() << interface.name() + " "+ entry.ip().toString() +" " + interface.hardwareAddress();
+
+                        NetworkDeviceModel *model = m_model->deviceByName(interface.name());
+
+                        if (!model) {
+                            model = new NetworkDeviceModel;
+                            model->setName(interface.name());
+                            model->setIp(entry.ip().toString());
+                            model->setHardwareAddress(interface.hardwareAddress());
+                            m_model->addDevice(model);
+                        } else {
+                            model->setName(interface.name());
+                            model->setIp(entry.ip().toString());
+                            model->setHardwareAddress(interface.hardwareAddress());
+                        }
+
+                        device = interface.name();
+                        if (m_rx) {
+                            m_rx->deleteLater();
+                            m_rx = nullptr;
+                        }
+                        if (m_tx) {
+                            m_tx->deleteLater();
+                            m_tx = nullptr;
+                        }
+                        m_rx = new QFile("/sys/class/net/" + interface.name() + "/statistics/rx_bytes");
+                        m_tx = new QFile("/sys/class/net/" + interface.name() + "/statistics/tx_bytes");
+                        break;
+                    }
+                }
             }
         }
